Create a separate file for Statement node
